moved copyCWDToId to a test_helpers file under test
