+++ include/CharBuf.h
@@ -107,16 +107,4 @@ int getIndexForChar(CharBuf buf, char c, int index, int direction);
  */
 CharBuf copyBytesFromBuf(CharBuf buf, int index);
 
-/**
- * <pre>
- * This function creates a new CharBuf that contains the contents of two character strings
- * </pre>
- *
- * @param[in] prefix The character string of the key_id without current working directory prefix (schema notation)
- * @param[in] postfix The character string of the key_id without current working directory postfix (file path)
- *
- * @return CharBuf copy of key_id with current working directory
- */
-CharBuf copyCWDToId(char *prefix, char *postfix);
-
 #endif /* INCLUDE_CHARBUF_H_ */
