+++ src/util/CharBuf.c
@@ -86,16 +86,3 @@ CharBuf copyBytesFromBuf(CharBuf buf, int index)
   memcpy(newBuf.chars, &buf.chars[index], newBuf.len);
   return (newBuf);
 }
-
-CharBuf copyCWDToId(char *prefix, char *postfix)
-{
-  CharBuf newBuf;
-  char cwd[100];
-
-  getcwd(cwd, sizeof(cwd));
-  newBuf = newCharBuf(strlen(prefix) + strlen(cwd) + strlen(postfix));
-  memcpy(newBuf.chars, prefix, strlen(prefix));
-  memcpy(&newBuf.chars[strlen(prefix)], cwd, strlen(cwd));
-  memcpy(&newBuf.chars[strlen(prefix) + strlen(cwd)], postfix, strlen(postfix));
-  return (newBuf);
-}
