+++ test/src/util/util_test_suite.c
@@ -3,6 +3,7 @@
  */
 
 #include "util_test_suite.h"
+#include "test_helper_functions.h"
 
 #include <stdio.h>
 #include <stdlib.h>
